Removed `send_ref_offset_value' which hasen't been used for awhile. Fixed various C compiler warnings raised by clang. Now the runtime compiles without warnings with -Wall and -Wextra, but there are still quite a bit of warnings if using -Weverything: * Change many types from numeral int/long to size_t which in many cases enable us to get rid of casts (that are always hiding some issues). * Change some numeral to `uint32' to match what is expected (but maybe size_t would be better, but this is a more complicated change). * Added new header `ewb_init.h' to provide a prototype to public routines of `ewb_init.c'. * Fixed missing declarations, unused macros, ... * Changed declaration of EIF_process_once_count to be size_t and avoid many warnings.
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@95924 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/C/ipc/ewb/eif_in.c
@@ -158,7 +158,7 @@ rt_public EIF_REFERENCE request_dispatch (Request rqst)
 				ptr += strlen (ptr) + 1;
 				sprintf (ptr, "0x%" EIF_POINTER_DISPLAY, (rt_uint_ptr) notif_info.st_data2);
 				ptr += strlen (ptr) + 1;
-				eif_string = makestr (string, ptr - string);
+				eif_string = makestr (string, (size_t) (ptr - string));
 				(notify_hdlr_set) (eif_access (notify_handler), eif_string);
 				return eif_access (notify_handler);
 			}
@@ -188,7 +188,7 @@ rt_public EIF_REFERENCE request_dispatch (Request rqst)
 				ptr += strlen (ptr) + 1;
 				sprintf (ptr, "%i", stop_info.st_exception);
 				ptr += strlen (ptr); /* terminating null so that (ptr - string) is the length */
-				eif_string = makestr (string, ptr - string);
+				eif_string = makestr (string, (size_t) (ptr - string));
 				(stopped_hdlr_set) (eif_access (stopped_handler), eif_string);
 				return eif_access (stopped_handler);
 			}
